@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Simorgh",
   "scripts": {
-    "amp:validate": "wait-on http://localhost:7080/news/articles/cl55zn0w0l4o.amp && amphtml-validator http://localhost:7080/news/articles/cl55zn0w0l4o.amp",
+    "amp:validate": "wait-on http://localhost:7080/news/articles/c6v11qzyv8po.amp && amphtml-validator http://localhost:7080/news/articles/c6v11qzyv8po.amp",
     "bbcA11y": "bbc-a11y",
     "build:ci": "export NODE_ENV=production && rm -rf build && npm run build:test && npm run build:live",
     "build": "rm -rf build && cp envConfig/local.env .env && NODE_ENV=production webpack",
@@ -119,10 +119,11 @@
     "path-to-regexp": "^2.4.0",
     "polyfill-crypto.getrandomvalues": "1.0.0",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0",
+    "react": "^16.8.6",
     "react-amphtml": "^3.0.1",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.8.6",
     "react-helmet": "^5.2.0",
+    "react-lazyload": "^2.5.0",
     "react-router-config": "^5.0.0",
     "react-router-dom": "^4.3.1",
     "speculate": "^1.7.4",
@@ -162,7 +163,7 @@
     "jest-styled-components": "^6.3.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^1.15.0",
-    "react-test-renderer": "^16.7.0",
+    "react-test-renderer": "^16.8.6",
     "react-testing-library": "^7.0.0",
     "snyk": "^1.129.0",
     "start-server-webpack-plugin": "^2.2.5",
@@ -205,7 +206,8 @@
     "prune": false,
     "requires": [
       "td-agent",
-      "cloudteam-td-agent-setup"
+      "cloudteam-td-agent-setup",
+      "dial-agent"
     ]
   }
 }
