@@ -1,95 +1,65 @@
-/* eslint-disable */
-/*
- * © Jordan Tart https://github.com/jtart
- * https://github.com/jtart/react-universal-app
- */
-import { Component } from 'react';
+import { useEffect, useState, useRef } from 'react';
 import { renderRoutes } from 'react-router-config';
 import { withRouter } from 'react-router-dom';
 import getRouteProps from '../../routes/getInitialData/utils/getRouteProps';
 
-import loadInitialData from '../../routes/loadInitialData';
+export const App = ({ routes, location, initialData, bbcOrigin }) => {
+  const { service, isAmp, id } = getRouteProps(routes, location.pathname);
 
-export class App extends Component {
-  constructor(props) {
-    super(props);
+  const [state, setState] = useState({
+    data: initialData,
+    service,
+    id,
+    isAmp,
+    loading: false,
+    error: null,
+  });
 
-    const { service, isAmp, id } = getRouteProps(
-      this.props.routes,
-      this.props.location.pathname,
-    );
+  const isInitialMount = useRef(true);
 
-    this.state = {
-      data: this.props.initialData,
-      service: service,
-      id: id,
-      isAmp: isAmp,
-      loading: false,
-      error: null,
-      loadInitialDataPromise: null,
-    };
-  }
+  useEffect(() => {
+    if (isInitialMount.current) {
+      isInitialMount.current = false;
+    } else {
+      // Only update on subsequent page renders
+      const {
+        service: nextService,
+        id: nextId,
+        isAmp: nextIsAmp,
+        route,
+        match,
+      } = getRouteProps(routes, location.pathname);
 
-  async componentDidUpdate({ location: prevLocation }) {
-    const { service, isAmp, id } = getRouteProps(
-      this.props.routes,
-      this.props.location.pathname,
-    );
-
-    if (this.props.location.pathname !== prevLocation.pathname) {
-      const initialData = loadInitialData(
-        this.props.location.pathname,
-        this.props.routes,
-      );
-
-      this.setState({
+      setState({
         data: null,
-        service: service,
-        id: id,
-        isAmp: isAmp,
+        service: nextService,
+        id: nextId,
+        isAmp: nextIsAmp,
         loading: true,
         error: null,
-        loadInitialDataPromise: initialData,
       });
-    }
 
-    if (this.state.loading) {
-      try {
-        const data = await this.state.loadInitialDataPromise;
-        this.setState({
-          data,
-          service: service,
-          id: id,
-          isAmp: isAmp,
-          loading: false,
-          error: null,
-          loadInitialDataPromise: null,
-        });
-      } catch (error) {
-        this.setState({
-          data: null,
-          service: service,
-          id: id,
-          isAmp: isAmp,
-          loading: false,
-          error,
-          loadInitialDataPromise: null,
-        });
-      }
+      const fetchData = async () => {
+        try {
+          const newData = await route.getInitialData(match.params);
+          setState(prevState => ({
+            ...prevState,
+            data: newData,
+            loading: false,
+          }));
+        } catch (error) {
+          setState(prevState => ({
+            ...prevState,
+            error,
+            loading: false,
+          }));
+        }
+      };
+      fetchData();
     }
-  }
+  }, [routes, location.pathname]);
 
-  render() {
-    return renderRoutes(this.props.routes, {
-      data: this.state.data,
-      service: this.state.service,
-      id: this.state.id,
-      isAmp: this.state.isAmp,
-      loading: this.state.loading,
-      error: this.state.error,
-      bbcOrigin: this.props.bbcOrigin,
-    });
-  }
-}
+  return renderRoutes(routes, { ...state, bbcOrigin });
+};
 
 export default withRouter(App);
