@@ -38,7 +38,25 @@ describeForLocalOnly('Article Body Tests', () => {
   });
 
   it('should have a visible image with a caption', () => {
-    visibleImageWithCaption(getElement('figure').eq(2));
+    let thirdFigure;
+
+    thirdFigure = getElement('figure').eq(2);
+
+    thirdFigure.within(() => {
+      const lazyLoadPlaceholder = getElement('div div');
+      lazyLoadPlaceholder.should('have.class', 'lazyload-placeholder');
+
+      const noscriptImg = getElement('noscript');
+      noscriptImg.contains('<img ');
+
+      cy.scrollTo('bottom', { duration: 200 });
+
+      const ImageContainer = getElement('div div');
+      ImageContainer.should('not.have.class', 'lazyload-placeholder');
+    });
+
+    thirdFigure = getElement('figure').eq(2);
+    visibleImageWithCaption(thirdFigure);
   });
 
   it('should render a title', () => {
