@@ -17,9 +17,9 @@ const config = {
     assetUrl: 'https://news.test.files.bbci.co.uk/include/articles/public',
     assetOrigin: 'https://news.test.files.bbci.co.uk',
     assets: {
-      news: 'cn7769kpk9mo',
-      newsThreeSubheadlines: 'cl55zn0w0l4o',
-      persian: 'cyddjz5058wo',
+      news: 'c0g992jmmkko',
+      newsThreeSubheadlines: 'c6v11qzyv8po',
+      persian: 'c4vlle3q337o',
       nonExistent: 'cxvxrj8tvppo',
     },
   },
@@ -29,9 +29,9 @@ const config = {
     assetUrl: 'http://localhost:7080',
     assetOrigin: 'http://localhost:7080',
     assets: {
-      news: 'cn7769kpk9mo',
-      newsThreeSubheadlines: 'cl55zn0w0l4o',
-      persian: 'cyddjz5058wo',
+      news: 'c0g992jmmkko',
+      newsThreeSubheadlines: 'c6v11qzyv8po',
+      persian: 'c4vlle3q337o',
       nonExistent: 'cxvxrj8tvppo',
     },
   },
