@@ -4,19 +4,18 @@
  */
 import React from 'react';
 import * as reactRouterConfig from 'react-router-config';
+import { mount } from 'enzyme';
+import { act } from 'react-dom/test-utils';
 import { App } from './App';
-import * as loadInitialData from '../../routes/loadInitialData';
 
 describe('App', () => {
   let wrapper;
-  let setStateSpy;
-  let loadInitialDataSpy;
   const initialData = { pageData: 'Some initial data' };
   const error = 'Error!';
   const match = { params: { service: 'news', amp: false } };
 
   const route = {
-    getInitialData: jest.fn().mockImplementation(() => Promise.reject(error)),
+    getInitialData: jest.fn(),
   };
 
   reactRouterConfig.matchRoutes = jest.fn().mockReturnValue([{ route, match }]);
@@ -26,8 +25,7 @@ describe('App', () => {
     .mockReturnValue(<h1>{initialData.pageData}</h1>);
 
   beforeAll(() => {
-    // eslint-disable-next-line no-undef
-    wrapper = shallow(
+    wrapper = mount(
       <App
         location={{ pathname: 'pathnameOne' }}
         routes={[]}
@@ -35,13 +33,11 @@ describe('App', () => {
         bbcOrigin="https://www.bbc.co.uk"
       />,
     );
-    setStateSpy = jest.spyOn(wrapper.instance(), 'setState');
-    loadInitialDataSpy = jest.spyOn(loadInitialData, 'default');
   });
 
   it('should return rendered routes', () => {
     expect.assertions(4);
-    expect(loadInitialDataSpy).not.toHaveBeenCalled();
+    expect(route.getInitialData).not.toHaveBeenCalled();
     expect(reactRouterConfig.renderRoutes).toHaveBeenCalledTimes(1);
     expect(reactRouterConfig.renderRoutes).toHaveBeenCalledWith([], {
       data: initialData,
@@ -57,50 +53,61 @@ describe('App', () => {
   describe('componentDidUpdate', () => {
     describe('same location', () => {
       it('should not call set state with new data', () => {
-        wrapper.setProps({ location: { pathname: 'pathnameOne' } });
+        reactRouterConfig.renderRoutes.mockClear();
 
-        expect.assertions(2);
-        expect(loadInitialDataSpy).not.toHaveBeenCalled();
-        expect(setStateSpy).not.toHaveBeenCalled();
+        act(() => {
+          wrapper.setProps({ location: { pathname: 'pathnameOne' } });
+
+          expect.assertions(2);
+          expect(route.getInitialData).not.toHaveBeenCalled();
+          expect(reactRouterConfig.renderRoutes).not.toHaveBeenCalled();
+        });
       });
     });
 
     describe('different location', () => {
-      afterEach(() => {
-        // clear `loadInitialData.default` mocks
+      beforeEach(() => {
+        // clear `route.getInitialData` and `reactRouterConfig.renderRoutes` mocks
         jest.clearAllMocks();
       });
       describe('rejected loadInitialData', () => {
         it('should set state to the error', async () => {
-          loadInitialData.default = jest
-            .fn()
-            .mockImplementation(() => Promise.reject(error));
+          route.getInitialData.mockImplementation(() => Promise.reject(error));
 
-          wrapper.setProps({ location: { pathname: 'pathnameThree' } });
+          act(() => {
+            wrapper.setProps({ location: { pathname: 'pathnameThree' } });
+          });
 
-          await loadInitialData.default;
+          await route.getInitialData;
 
           expect.assertions(2);
 
-          // start data fetch and set loading to true
-          expect(setStateSpy).toHaveBeenNthCalledWith(1, {
-            data: null,
-            service: 'news',
-            isAmp: false,
-            error: null,
-            loadInitialDataPromise: expect.any(Promise),
-            loading: true,
-          });
+          expect(reactRouterConfig.renderRoutes).toHaveBeenNthCalledWith(
+            2,
+            [],
+            {
+              data: null,
+              service: 'news',
+              isAmp: false,
+              error: null,
+              loading: true,
+              bbcOrigin: 'https://www.bbc.co.uk',
+            },
+          );
 
           // data fetch promise rejected, set data to null, loading to false and set error
-          expect(setStateSpy).toHaveBeenNthCalledWith(2, {
-            data: null,
-            isAmp: false,
-            error,
-            loadInitialDataPromise: null,
-            loading: false,
-            service: 'news',
-          });
+          expect(reactRouterConfig.renderRoutes).toHaveBeenNthCalledWith(
+            3,
+            [],
+            {
+              data: null,
+              service: 'news',
+              isAmp: false,
+              error,
+              loading: false,
+              bbcOrigin: 'https://www.bbc.co.uk',
+            },
+          );
         });
       });
 
@@ -109,37 +116,43 @@ describe('App', () => {
           const pathname = 'pathnameFour';
           const data = 'Really cool data';
 
-          loadInitialData.default = jest
-            .fn()
-            .mockImplementation(async () => data);
+          route.getInitialData.mockImplementation(() => Promise.resolve(data));
 
           wrapper.setProps({ location: { pathname } });
 
-          await loadInitialData.default;
+          await route.getInitialData;
 
           expect.assertions(3);
 
-          expect(loadInitialData.default).toHaveBeenCalledWith(pathname, []);
+          expect(route.getInitialData).toHaveBeenCalledWith(match.params);
 
           // start data fetch and set loading to true
-          expect(setStateSpy).toHaveBeenNthCalledWith(1, {
-            data: null,
-            error: null,
-            isAmp: false,
-            loading: true,
-            loadInitialDataPromise: expect.any(Promise),
-            service: 'news',
-          });
+          expect(reactRouterConfig.renderRoutes).toHaveBeenNthCalledWith(
+            2,
+            [],
+            {
+              data: null,
+              service: 'news',
+              isAmp: false,
+              error: null,
+              loading: true,
+              bbcOrigin: 'https://www.bbc.co.uk',
+            },
+          );
 
           // data fetch promise resolved, set data to fetched data and loading to false
-          expect(setStateSpy).toHaveBeenNthCalledWith(2, {
-            data,
-            error: null,
-            isAmp: false,
-            loading: false,
-            loadInitialDataPromise: null,
-            service: 'news',
-          });
+          expect(reactRouterConfig.renderRoutes).toHaveBeenNthCalledWith(
+            3,
+            [],
+            {
+              data,
+              service: 'news',
+              isAmp: false,
+              error: null,
+              loading: false,
+              bbcOrigin: 'https://www.bbc.co.uk',
+            },
+          );
         });
       });
     });
