@@ -1,11 +1,26 @@
-import React, { useContext } from 'react';
-import { string, number, objectOf, any } from 'prop-types';
+import React, { useContext, Fragment } from 'react';
+import { string, number, objectOf, any, bool } from 'prop-types';
 import Figure from '@bbc/psammead-figure';
 import Image, { AmpImg } from '@bbc/psammead-image';
 import ImagePlaceholder from '@bbc/psammead-image-placeholder';
+import LazyLoad from 'react-lazyload';
 import Copyright from '../Copyright';
 import Caption from '../Caption';
 import { RequestContext } from '../../contexts/RequestContext';
+
+const LAZYLOAD_OFFSET = 250; // amount of pixels below the viewport to begin loading the image
+
+const renderImage = (imageToRender, lazyLoad) =>
+  lazyLoad ? (
+    <Fragment>
+      <LazyLoad offset={LAZYLOAD_OFFSET} once>
+        {imageToRender}
+      </LazyLoad>
+      <noscript>{imageToRender}</noscript>
+    </Fragment>
+  ) : (
+    imageToRender
+  );
 
 const renderCopyright = copyright =>
   copyright ? <Copyright>{copyright}</Copyright> : null;
@@ -17,6 +32,7 @@ const FigureContainer = ({
   alt,
   copyright,
   captionBlock,
+  lazyLoad,
   ratio,
   src,
   height,
@@ -24,6 +40,8 @@ const FigureContainer = ({
   type,
 }) => {
   const { platform } = useContext(RequestContext);
+  const imageToRender = <Image alt={alt} src={src} width={width} />;
+
   return (
     <Figure>
       <ImagePlaceholder ratio={ratio}>
@@ -37,7 +55,7 @@ const FigureContainer = ({
             width={width}
           />
         ) : (
-          <Image alt={alt} src={src} width={width} />
+          renderImage(imageToRender, lazyLoad)
         )}
         {renderCopyright(copyright)}
       </ImagePlaceholder>
@@ -50,17 +68,19 @@ FigureContainer.propTypes = {
   alt: string.isRequired,
   captionBlock: objectOf(any),
   copyright: string,
+  height: number,
+  lazyLoad: bool,
   ratio: number.isRequired,
   src: string.isRequired,
-  height: number,
-  width: number.isRequired,
   type: string,
+  width: number.isRequired,
 };
 
 FigureContainer.defaultProps = {
   copyright: null,
   captionBlock: null,
   height: null,
+  lazyLoad: false,
   type: '',
 };
 
