@@ -223,7 +223,7 @@ describe('Metadata Container', () => {
           dotComOrigin,
           'canonical',
           articleDataPersian,
-          'cyddjz5058wo',
+          'c4vlle3q337o',
         ),
       );
 
@@ -234,8 +234,8 @@ describe('Metadata Container', () => {
         metadataProps(
           false,
           [],
-          'https://www.bbc.com/persian/articles/cyddjz5058wo.amp',
-          'https://www.bbc.com/persian/articles/cyddjz5058wo',
+          'https://www.bbc.com/persian/articles/c4vlle3q337o.amp',
+          'https://www.bbc.com/persian/articles/c4vlle3q337o',
           'خلاصه مقاله',
           'fa',
           [],
@@ -246,7 +246,7 @@ describe('Metadata Container', () => {
       expect(Wrapper.find(LinkedData).props()).toEqual(
         linkedDataProps(
           'BBC News فارسی',
-          'https://www.bbc.com/persian/articles/cyddjz5058wo',
+          'https://www.bbc.com/persian/articles/c4vlle3q337o',
           'Persian',
           'https://news.files.bbci.co.uk/ws/img/logos/og/persian.png',
           'سرصفحه مقاله',
@@ -261,7 +261,7 @@ describe('Metadata Container', () => {
           dotCoDotUKOrigin,
           'amp',
           articleDataPersian,
-          'cyddjz5058wo',
+          'c4vlle3q337o',
         ),
       );
 
@@ -272,8 +272,8 @@ describe('Metadata Container', () => {
         metadataProps(
           true,
           [],
-          'https://www.bbc.co.uk/persian/articles/cyddjz5058wo.amp',
-          'https://www.bbc.co.uk/persian/articles/cyddjz5058wo',
+          'https://www.bbc.co.uk/persian/articles/c4vlle3q337o.amp',
+          'https://www.bbc.co.uk/persian/articles/c4vlle3q337o',
           'خلاصه مقاله',
           'fa',
           [],
@@ -284,7 +284,7 @@ describe('Metadata Container', () => {
       expect(Wrapper.find(LinkedData).props()).toEqual(
         linkedDataProps(
           'BBC News فارسی',
-          'https://www.bbc.co.uk/persian/articles/cyddjz5058wo',
+          'https://www.bbc.co.uk/persian/articles/c4vlle3q337o',
           'Persian',
           'https://news.files.bbci.co.uk/ws/img/logos/og/persian.png',
           'سرصفحه مقاله',
