@@ -2,9 +2,7 @@ import onClient from '../../helpers/onClient';
 import getBaseUrl from './utils/getBaseUrl';
 import fetchData from './utils/fetchData';
 
-const getArticleInitialData = async ({ match }) => {
-  const { id, service } = match.params;
-
+const getArticleInitialData = async ({ id, service }) => {
   const baseUrl = onClient()
     ? getBaseUrl(window.location.origin)
     : process.env.SIMORGH_BASE_URL;
