@@ -3,8 +3,11 @@ import request from 'supertest';
 import * as reactDomServer from 'react-dom/server';
 import * as styledComponents from 'styled-components';
 import dotenv from 'dotenv';
-import loadInitialData from '../app/routes/loadInitialData';
+import getRouteProps from '../app/routes/getInitialData/utils/getRouteProps';
 import Document from './Document/component';
+import getDials from './getDials';
+
+jest.mock('./getDials');
 
 // mimic the logic in `src/index.js` which imports the `server/index.jsx`
 dotenv.config({ path: './envConfig/local.env' });
@@ -33,7 +36,29 @@ jest.mock('react-helmet', () => ({
   },
 }));
 
-jest.mock('../app/routes/loadInitialData');
+jest.mock('../app/routes/getInitialData/utils/getRouteProps');
+
+const mockRouteProps = (
+  id,
+  service,
+  isAmp,
+  dataResponse,
+  responseType = 'resolve',
+) => {
+  const getInitialData =
+    responseType === 'reject'
+      ? jest.fn().mockRejectedValueOnce(dataResponse)
+      : jest.fn().mockResolvedValueOnce(dataResponse);
+
+  getRouteProps.mockReturnValue({
+    isAmp,
+    service,
+    route: { getInitialData },
+    match: {
+      params: { id, service },
+    },
+  });
+};
 
 styledComponents.ServerStyleSheet = jest.fn().mockImplementation(() => ({
   collectStyles: jest.fn().mockReturnValue(<h1>Mock app</h1>),
@@ -64,7 +89,7 @@ describe('Server', () => {
   describe('Data', () => {
     describe('for articles', () => {
       it('should respond with JSON', async () => {
-        const { body } = await makeRequest('/news/articles/cn7769kpk9mo.json');
+        const { body } = await makeRequest('/news/articles/c0g992jmmkko.json');
         expect(body).toEqual(
           expect.objectContaining({ content: expect.any(Object) }),
         );
@@ -140,15 +165,20 @@ describe('Server', () => {
 
     describe('Successful render', () => {
       describe('200 status code', () => {
+        const id = 'c0000000001o';
+        const service = 'news';
+        const isAmp = false;
+
         beforeEach(() => {
-          loadInitialData.mockImplementationOnce(() =>
-            Promise.resolve(successDataResponse),
-          );
+          mockRouteProps(id, service, isAmp, successDataResponse);
         });
 
         it('should respond with rendered data', async () => {
+          const dials = { dial: 'value' };
+          getDials.mockResolvedValue(dials);
+
           const { text, status } = await makeRequest(
-            '/news/articles/c0000000001o',
+            `/${service}/articles/${id}`,
           );
 
           expect(status).toBe(200);
@@ -168,9 +198,10 @@ describe('Server', () => {
               ]}
               data={successDataResponse}
               helmet={{ head: 'tags' }}
-              isAmp={false}
-              service="news"
+              isAmp={isAmp}
+              service={service}
               styleTags={<style />}
+              dials={dials}
             />,
           );
 
@@ -178,18 +209,26 @@ describe('Server', () => {
             '<!doctype html><html><body><h1>Mock app</h1></body></html>',
           );
         });
+
+        it('should respond successfully even if dials fetch fails', async () => {
+          getDials.mockRejectedValue(new Error('Fetch fail'));
+
+          const { status } = await makeRequest('/news/articles/c0000000001o');
+          expect(status).toBe(200);
+        });
       });
 
       describe('404 status code', () => {
+        const id = 'c0000000001o';
+        const service = 'news';
+
         beforeEach(() => {
-          loadInitialData.mockImplementationOnce(() =>
-            Promise.resolve(notFoundDataResponse),
-          );
+          mockRouteProps(id, service, false, notFoundDataResponse);
         });
 
         it('should respond with a rendered 404', async () => {
           const { status, text } = await makeRequest(
-            '/news/articles/c0000000001o',
+            `/${service}/articles/${id}`,
           );
           expect(status).toBe(404);
           expect(text).toEqual(
@@ -200,15 +239,16 @@ describe('Server', () => {
     });
 
     describe('Unknown error within universal-react-app or its dependencies', () => {
+      const id = 'c0000000001o';
+      const service = 'news';
+
       beforeEach(() => {
-        loadInitialData.mockImplementationOnce(() =>
-          Promise.reject(Error('Error!')),
-        );
+        mockRouteProps(id, service, false, Error('Error!'), 'reject');
       });
 
       it('should respond with a 500', async () => {
         const { status, text } = await makeRequest(
-          '/news/articles/c0000000001o',
+          `/${service}/articles/${id}`,
         );
         expect(status).toEqual(500);
         expect(text).toEqual('Error!');
